Travis problem and new test code for Event()
Travis uses go1.12 and my testing is using go1.14.6 and the compilers do not initialise the "test.v" flag in the same order which stopped my new test code that more fully tests Event() did not end up calling the eventWithOptionsCheck().
When adjustments made to fix this, another problem came to light with the existing code which called eventWithOptionsCheck() directly and that works, BUT calling Event() caused a panic.
The fix is to add a check to see if what is passed to eventWithOptionsCheck() is a pointer.