Namespace defaults to last element of path given as os.Args[0] (v1)
Namespace defaults to path.Base(os.Args[0]), i.e. the last element
of the path used to invoke the binary, rather than plain os.Args[0].

This change needed to avoid situtations where an app is using multiple
versions of log.go, but only one version has the Namespace var set
(generally the one that is used in main.go), and the other version(s)
of log.go use the default os.Args[0]. This has resulted in apps logging
to different namespaces, e.g. MyApp and ./MyApp, depending on which
version of log.go is doing the logging.