Bring in custom event functions from PR done before joining ONS
There is a lot more to do to bring the old code up to date with changes made to log lib after 3 years, and get it working.
There are 3 custom events that:
Reduce dp-frontend-router's HOT-PATH log.Event() memory allocations by up to 31%
Two of the custom log event handlers are in the log library, so improve all apps and one of them is placed in dp-frontend-router on its HOT happy path.